ref: a few code refactoring
