--- .env
@@ -1,2 +1,0 @@
-REACT_APP_TEST = 'test env'
-REACT_APP_WEBPACK_TEST = 'webpack test'
