+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 dist
 dist/*LICENSE.txt
 yarn-error.log
+.env
