+++ config/webpack/buildPlugins.ts
@@ -2,9 +2,19 @@ import webpack from 'webpack';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import MiniCssExtractPlugin from 'mini-css-extract-plugin';
 import { BuildOptions } from './types';
+import dotenv from 'dotenv';
 
 export function buildPlugins(options: BuildOptions): webpack.WebpackPluginInstance[] {
   const { paths } = options;
+  const env = dotenv.config().parsed;
+
+  const envKeys = Object.keys(env).reduce(
+    (prev, next) => {
+      prev[`process.env.${next}`] = JSON.stringify(env[next]);
+      return prev;
+    },
+    {} as Record<string, string>,
+  );
 
   return [
     new webpack.ProgressPlugin(),
@@ -15,9 +25,7 @@ export function buildPlugins(options: BuildOptions): webpack.WebpackPluginInstan
       template: paths.html,
       favicon: paths.favicon,
     }),
-    new webpack.DefinePlugin({
-      // 'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
-      'process.env.WEBPACK_TEST': JSON.stringify(process.env.WEBPACK_TEST),
-    }),
+
+    new webpack.DefinePlugin(envKeys),
   ];
 }
