+++ package.json
@@ -48,6 +48,7 @@
     "chai-colors": "^1.0.1",
     "css-loader": "^6.8.1",
     "cypress": "^12.17.4",
+    "dotenv": "^16.3.1",
     "eslint": "^8.44.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-react": "^7.32.2",
