+++ src/components/App/loader/index.scss
@@ -1,4 +1,6 @@
-.app {
+.loader {
+  justify-content: center;
+  align-items: center;
   min-height: 100vh;
   background-color: var(--bg);
   transition: var(--transition);
