+++ src/components/banks/search/option/index.scss
@@ -1,0 +1,10 @@
+@import '@styles/variables.scss';
+@import '@styles/templates.scss';
+
+.option {
+  border-bottom: 1px solid var(--gray2);
+  padding: 10px 0px;
+  color: var(--banks-text);
+  @extend %buttonHover;
+  @extend %buttonActive;
+}
