+++ src/components/currencies/index.scss
@@ -28,121 +28,3 @@
     justify-self: center;
   }
 }
-
-.currency {
-  display: flex;
-  width: 100%;
-  padding: 30px 32px 30px 32px;
-  align-items: center;
-  gap: 30px;
-  border-radius: 8px;
-  border: 1px solid var(--currency-card-border);
-  background: var(--currency-card-bg);
-  transition: var(--transition);
-
-  @extend %buttonHover;
-  @extend %buttonActive;
-
-  &:hover {
-    cursor: pointer;
-  }
-
-  @media (max-width: 1100px) {
-    padding: 30px;
-  }
-
-  @media (max-width: 768px) {
-    padding: 12px;
-    max-width: none;
-  }
-
-  &__icon {
-    width: 80px;
-    height: 80px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 30px;
-
-    @media (max-width: 1100px) {
-      width: 50px;
-      height: 50px;
-      font-size: 18px;
-    }
-
-    @media (max-width: 768px) {
-      width: 30px;
-      height: 30px;
-      font-size: 12px;
-    }
-  }
-
-  &__content {
-    display: flex;
-    flex-direction: column;
-    row-gap: 5px;
-  }
-
-  &__title {
-    color: var(--currency-card-title);
-    font-size: 35px;
-    font-weight: 400;
-    line-height: 141%;
-    transition: var(--transition);
-
-    @media (max-width: 1100px) {
-      font-size: 16px;
-    }
-
-    @media (max-width: 768px) {
-      font-size: 14px;
-    }
-  }
-
-  &__price {
-    color: var(--currency-card-cost);
-    transition: var(--transition);
-    font-size: 32px;
-    font-weight: 300;
-    line-height: 128%;
-
-    @media (max-width: 1100px) {
-      font-size: 16px;
-    }
-
-    @media (max-width: 768px) {
-      font-size: 12px;
-    }
-  }
-
-  &-1 {
-    background: var(--currency-symbol-bg-1);
-  }
-  &-2 {
-    background: var(--currency-symbol-bg-2);
-  }
-  &-3 {
-    background: var(--currency-symbol-bg-3);
-  }
-  &-4 {
-    background: var(--currency-symbol-bg-4);
-  }
-  &-5 {
-    background: var(--currency-symbol-bg-5);
-  }
-  &-6 {
-    background: var(--currency-symbol-bg-6);
-  }
-  &-7 {
-    background: var(--currency-symbol-bg-7);
-  }
-  &-8 {
-    background: var(--currency-symbol-bg-8);
-  }
-  &-9 {
-    background: var(--currency-symbol-bg-9);
-  }
-  &-10 {
-    background: var(--currency-symbol-bg-10);
-  }
-}
