+++ src/components/footer/index.scss
@@ -95,91 +95,3 @@
     }
   }
 }
-
-.navigation {
-  flex: 1 1 auto;
-  display: flex;
-  column-gap: 30px;
-  justify-content: space-between;
-
-  @media (max-width: 600px) {
-    flex-direction: column;
-  }
-
-  &__column {
-    @media (max-width: 600px) {
-      display: flex;
-      flex-direction: column;
-      padding: 15px 0px;
-      border-bottom: 1px solid var(--gray2);
-
-      &.active {
-        .navigation__links {
-          height: 70px;
-        }
-
-        .navigation__title {
-          &:before {
-            transform: rotate(180deg);
-            transition: var(--transition);
-          }
-        }
-      }
-    }
-  }
-
-  &__title {
-    font-size: 28px;
-    font-weight: 400;
-    line-height: 150%;
-    color: var(--footer-title);
-    transition: var(--transition);
-    margin-bottom: 40px;
-
-    @media (max-width: 600px) {
-      font-size: 20px;
-      font-weight: 400;
-      letter-spacing: -0.24px;
-      margin-bottom: 0px;
-
-      position: relative;
-
-      &::before {
-        content: '';
-        position: absolute;
-        right: 0;
-        top: 5px;
-        height: 20px;
-        width: 20px;
-        background: url('../../assets//icons/arrow.svg') right center no-repeat;
-        transition: var(--transition);
-      }
-    }
-  }
-
-  &__links {
-    display: flex;
-    flex-direction: column;
-    font-size: 24px;
-    font-weight: 400;
-    row-gap: 18px;
-    color: var(--footer-links);
-    transition: height 0.3s ease;
-
-    @media (max-width: 600px) {
-      height: 0px;
-      overflow: hidden;
-    }
-  }
-
-  &__link {
-    & a {
-      @extend %buttonHover;
-      @extend %buttonActive;
-
-      @media (max-width: 600px) {
-        font-size: 20px;
-      }
-    }
-  }
-}
