+++ src/components/header/index.tsx
@@ -6,6 +6,8 @@ import { INavLink } from '@constants/types';
 import { FC } from 'react';
 import { Link } from 'react-router-dom';
 
+import { NavLink } from './nav-link';
+
 export const Header: FC = () => {
   const pagesList: INavLink[] = [
     { name: 'Home', route: '/' },
@@ -30,16 +32,5 @@ export const Header: FC = () => {
         </nav>
       </div>
     </header>
-  );
-};
-
-const NavLink: FC<INavLink> = ({ name, route }) => {
-  const linkName = name.split(' ');
-  const testID = linkName[linkName.length - 1].toLowerCase() + '-route';
-
-  return (
-    <li className='header__link' data-testid={testID}>
-      <Link to={route}>{name}</Link>
-    </li>
   );
 };
