+++ src/components/header/nav-link/index.scss
@@ -1,0 +1,15 @@
+@import '@styles/variables.scss';
+@import '@styles/templates.scss';
+
+.header {
+  &__link {
+    color: var(--navLinkColor);
+    font-weight: 300;
+    line-height: calc(40 / 20 * 100%);
+    user-select: none;
+    white-space: nowrap;
+
+    @extend %buttonHover;
+    @extend %buttonActive;
+  }
+}
