+++ src/components/modal/index.scss
@@ -8,7 +8,6 @@
   bottom: 0;
   right: 0;
   background: rgba($color: #000, $alpha: 0.7);
-  // opacity: 0.85;
 
   display: flex;
   justify-content: center;
@@ -42,65 +41,5 @@
     font-size: 20px;
     text-align: center;
     margin-bottom: 30px;
-  }
-}
-
-.select {
-  &__current {
-    display: flex;
-    flex-direction: column;
-    padding: 15px 0px;
-    border-bottom: 1px solid var(--gray2);
-    position: relative;
-    width: 100%;
-    padding-left: 10px;
-    transition: var(--transition);
-
-    &:hover {
-      transition: var(--transition);
-      @extend %buttonHover;
-    }
-
-    &::before {
-      content: '';
-      position: absolute;
-      right: 0;
-      height: 20px;
-      width: 20px;
-      background: url('../../assets//icons/arrow.svg') right center no-repeat;
-      transition: var(--transition);
-    }
-
-    &.active::before {
-      transform: rotate(180deg);
-      transition: var(--transition);
-    }
-  }
-
-  &__menu {
-    display: flex;
-    flex-direction: column;
-    background: #333;
-    color: white;
-    max-height: 230px;
-    overflow: auto;
-    height: 0px;
-    transition: var(--transition);
-    padding: 0px 10px 0px 10px;
-
-    &.active {
-      height: 230px;
-      transition: var(--transition);
-    }
-  }
-
-  &__option {
-    border-bottom: 1px solid var(--gray2);
-    padding: 10px 0px;
-
-    &:hover {
-      transition: var(--transition);
-      @extend %buttonHover;
-    }
   }
 }
