+++ src/components/modal/option/index.scss
@@ -1,0 +1,12 @@
+@import '@styles/variables.scss';
+@import '@styles/templates.scss';
+
+.option {
+  border-bottom: 1px solid var(--gray2);
+  padding: 10px 0px;
+
+  &:hover {
+    transition: var(--transition);
+    @extend %buttonHover;
+  }
+}
