+++ src/components/timeline/index.scss
@@ -14,108 +14,3 @@
   display: flex;
   flex-direction: column;
 }
-
-.current {
-  font-size: 25px;
-  width: 100%;
-  max-width: 300px;
-  line-height: normal;
-  margin-bottom: 20px;
-
-  &__title {
-    font-weight: 500;
-    position: relative;
-    color: var(--timeline-cost-text);
-
-    &:hover {
-      cursor: pointer;
-    }
-
-    &::before {
-      content: '';
-      position: absolute;
-      right: 0;
-      top: 30%;
-      height: 20px;
-      width: 20px;
-      background: url('../../assets/icons/arrow.svg') right center no-repeat;
-      transition: var(--transition);
-    }
-
-    &.active::before {
-      transform: rotate(180deg);
-      transition: var(--transition);
-    }
-  }
-
-  &__options {
-    position: absolute;
-    width: 100%;
-    max-width: 300px;
-    background-color: var(--black);
-    overflow: auto;
-    z-index: 150;
-  }
-
-  &__option {
-    font-size: 18px;
-    border-bottom: 1px solid var(--gray2);
-    padding: 10px 0px;
-
-    &:hover {
-      transition: var(--transition);
-      @extend %buttonHover;
-    }
-  }
-}
-
-.cost {
-  display: flex;
-  flex-direction: column;
-  row-gap: 10px;
-
-  &__title {
-    color: var(--timeline-cost-text);
-  }
-
-  &__body {
-    display: flex;
-    flex-direction: column;
-    row-gap: 30px;
-    margin-bottom: 20px;
-  }
-
-  &__input {
-    width: 100%;
-    height: 30px;
-    background-color: var(--timeline-select-bg);
-    color: var(--timeline-select-text);
-    padding-left: 10px;
-  }
-
-  &__content {
-    display: grid;
-    grid-template-columns: repeat(10, 1fr);
-    gap: 20px;
-    margin-bottom: 20px;
-
-    @media (max-width: 992px) {
-      grid-template-columns: repeat(5, 1fr);
-    }
-
-    @media (max-width: 480px) {
-      grid-template-columns: repeat(3, 1fr);
-    }
-  }
-
-  &__button {
-    height: 60px;
-    width: 300px;
-    align-self: center;
-    background: var(--green);
-    color: var(--white);
-
-    @extend %buttonHover;
-    @extend %buttonActive;
-  }
-}
