+++ src/pages/card/index.tsx
@@ -7,29 +7,13 @@ import { FC } from 'react';
 
 const CardPage: FC = () => {
   return (
-    // <section className='card' data-testid='card-page'>
     <div className='card__container'>
       <div className='card__body'>
         <Banner />
         <Banks />
       </div>
     </div>
-    // </section>
   );
 };
 
-// export default CardPage;
-
 export default withSection(CardPage, 'card', 'card-page');
-
-/*
-
- <section className={baseClassName} data-testid={testid}>
-          <div className={`${baseClassName}__container container`}>
-            <div className={`${baseClassName}__body`}>
-              <Banner />
-      <Banks />
-            </div>
-          </div>
-        </section>
-*/
