+++ src/store/index.ts
@@ -12,6 +12,3 @@ export const store = configureStore({
 });
 
 export type RootState = ReturnType<typeof store.getState>;
-
-// [movieApi.reducerPath]: movieApi.reducer,
-// },
